requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
psycopg2-binary
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
